+++ rust/src/ast/error.rs
@@ -1,0 +1,9 @@
+use thiserror::Error;
+
+pub type Result<T> = std::result::Result<T, Error>;
+
+#[derive(Debug, Error)]
+pub enum Error {
+    #[error("")]
+    DowncastFailed,
+}
