+++ rust/src/lexer/error.rs
@@ -1,0 +1,6 @@
+use thiserror::Error;
+
+pub type Result<T> = std::result::Result<T, Error>;
+
+#[derive(Debug, Error)]
+pub enum Error {}
