+++ rust/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod ast;
+pub mod lexer;
+pub mod parser;
+
+pub use lexer::Lexer;
+pub use parser::Parser;
